+++ www/views/Extjs3/javascript/form/HtmlEditor.js
@@ -8,6 +8,7 @@ GO.form.HtmlEditor = function (config) {
 	Ext.applyIf(config, {
 		border: false,
 		enableFont: false,
+		headingsMenu: true,
 		style: GO.settings.html_editor_font
 	});
 
@@ -25,6 +26,7 @@ GO.form.HtmlEditor = function (config) {
 	var ssScriptPlugin = new Ext.ux.form.HtmlEditor.SubSuperScript();
 	var rmFormatPlugin = new Ext.ux.form.HtmlEditor.RemoveFormat();
 
+
 	if (GO.settings.pspellSupport)
 		config.plugins.unshift(spellcheckInsertPlugin);
 
@@ -36,6 +38,11 @@ GO.form.HtmlEditor = function (config) {
 					ssScriptPlugin
 					);
 
+	if(config.headingsMenu) {
+		var headingMenu = new Ext.ux.form.HtmlEditor.HeadingMenu();
+		config.plugins.unshift(headingMenu);
+	}
+
 	GO.form.HtmlEditor.superclass.constructor.call(this, config);
 };
 
@@ -44,6 +51,8 @@ Ext.extend(GO.form.HtmlEditor, Ext.form.HtmlEditor, {
 	iframePad:dp(8),
 	
 	hideToolbar: false,
+
+	headingsMenu: true,
 
 	initComponent: function() {
 		GO.form.HtmlEditor.superclass.initComponent.apply(this);
@@ -403,7 +412,10 @@ Ext.extend(GO.form.HtmlEditor, Ext.form.HtmlEditor, {
 			try {
 				this.execCmd('useCSS', true);
 				this.execCmd('styleWithCSS', false);
+				this.execCmd('insertBrOnReturn', false);
+				this.execCmd('enableObjectResizing', true);
 			} catch (e) {
+				console.warn(e);
 			}
 		}
 		this.fireEvent('activate', this);
@@ -484,35 +496,30 @@ Ext.extend(GO.form.HtmlEditor, Ext.form.HtmlEditor, {
 					//                    }
 				}
 			};
-		} else if (Ext.isOpera) {
-			return function (e) {
-				var k = e.getKey();
-				if (k == e.TAB) {
-					e.stopEvent();
-					this.win.focus();
-					this.execCmd('InsertHTML', '&nbsp;&nbsp;&nbsp;&nbsp;');
-					this.deferFocus();
-				}
-			};
 		} else if (Ext.isWebKit) {
 			return function (e) {
-				var k = e.getKey();
+				var k = e.getKey(), doc = this.getDoc();
 				if (k == e.TAB) {
 					e.stopEvent();
 					this.execCmd('InsertText', '\t');
 					this.deferFocus();
-				}/* Testen in latest chrome and safari with lists else if (k == e.ENTER) {
-					e.stopEvent();
-					var doc = this.getDoc();
-					if (doc.queryCommandState('insertorderedlist') ||
-									doc.queryCommandState('insertunorderedlist')) {
-						this.execCmd('InsertHTML', '</li><br /><li>');
-					} else {
-						this.execCmd('InsertHtml', '<br />&nbsp;');
-						this.execCmd('delete');
-					}
-					this.deferFocus();
-				}*/
+				}else if(k == e.ENTER){
+					// if (doc.queryCommandState('insertorderedlist') || doc.queryCommandState('insertunorderedlist')) {
+					// 	return;
+					// }
+					// e.stopEvent();
+					//
+					//
+					// //make sure last child is a br otherwise it will go wrong!
+					// console.warn(doc.lastElementChild.tagName.toLowerCase());
+					// if(!doc.lastElementChild.tagName.toLowerCase() == "br") {
+					// 	console.warn("added br")
+					// 	doc.appendChild(doc.createElement("br"));
+					// }
+					//
+					// this.execCmd('InsertHtml','<br />');
+					// this.deferFocus();
+				}
 			};
 		}
 	}(),